split out pads to platform independant classes

+++ src/engine/core/String.h
@@ -47,7 +47,7 @@ public:
 
 public:
 
-  unsigned int size() const;
+  size_t size() const;
   
 public:
